# openapi.json
{
  "components": {
    "securitySchemes": {
      "sessao": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Gerado por `go run ./cmd/openapi`. As formas das respostas vêm dos golden files, que por sua vez vêm de respostas verdadeiras — não há terceira cópia escrita à mão.",
    "title": "Airo API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/training/today": {
      "get": {
        "operationId": "getTrainingToday",
        "responses": {
          "200": {
            "description": "Feito."
          },
          "401": {
            "description": "Sessão inválida ou expirada."
          }
        },
        "security": [
          {
            "sessao": []
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://airo-api.savanapoint.com"
    },
    {
      "description": "ensaio local",
      "url": "http://localhost:8090"
    }
  ]
}